Fix gh-9454 by replacing `int` -> `long` in bdtr.
Replacing `int` with `long` extends the range for which the function
returns valid results instead of NaN.

Alternatives considered:

1.  Use fixed-width types (`npy_int64`, `int64_t`, etc).  This may be
    better practice, but these types are not supported in
    `generate_ufuncs.pyx`, so using them would require a bigger
    refactoring (and discussion with developers).
2.  Use `double`.  Although bdtr has an integer signature, internally
    it calls `incbet`, which casts the integers to doubles.  For
    backwards compatibility we'd have to truncate the doubles to the
    nearest integer, but we would avoid imposing an arbitrary cutoff
    on the size of `n` and `k`.  The drawback is that the signature
    of the function no longer makes it clear that the input arguments
    will be truncated to integers.

+++ scipy/special/_legacy.pxd
@@ -60,7 +60,7 @@ cdef inline double bdtr_unsafe(double k, double n, double p) nogil:
     if npy_isnan(k) or npy_isnan(n):
         return nan
     _legacy_cast_check("bdtr", k, n)
-    return bdtr(<int>k, <int>n, p)
+    return bdtr(<long>k, <long>n, p)
 
 cdef inline double bdtri_unsafe(double k, double n, double y) nogil:
     if npy_isnan(k) or npy_isnan(n):
